+++ src/Govt.php
@@ -6,6 +6,65 @@ use JobBrander\Jobs\Client\Collection;
 class Govt extends AbstractProvider
 {
     /**
+     * Map of setter methods to query parameters
+     *
+     * @var array
+     */
+    protected $queryMap = [
+        'setCount' => 'size',
+        'setFrom' => 'from',
+        'setHl' => 'hl',
+        'setKeyword' => 'query',
+        'setLatLon' => 'lat_lon',
+        'setOrganizationIds' => 'organization_ids',
+        'setQuery' => 'query',
+        'setSize' => 'size',
+        'setTags' => 'tags',
+    ];
+
+    /**
+     * Current api query parameters
+     *
+     * @var array
+     */
+    protected $queryParams = [
+        'from' => null,
+        'hl' => null,
+        'lat_lon' => null,
+        'organization_ids' => null,
+        'query' => null,
+        'size' => null,
+        'tags' => null,
+    ];
+
+    /**
+     * Create new Govt jobs client.
+     *
+     * @param array $parameters
+     */
+    public function __construct($parameters = [])
+    {
+        parent::__construct($parameters);
+        array_walk($parameters, [$this, 'updateQuery']);
+    }
+
+    /**
+     * Magic method to handle get and set methods for properties
+     *
+     * @param  string $method
+     * @param  array  $parameters
+     *
+     * @return mixed
+     */
+    public function __call($method, $parameters)
+    {
+        if (isset($this->queryMap[$method], $parameters[0])) {
+            $this->updateQuery($parameters[0], $this->queryMap[$method]);
+        }
+        return parent::__call($method, $parameters);
+    }
+
+    /**
      * Returns the standardized job object
      *
      * @param array $payload
@@ -103,12 +162,20 @@ class Govt extends AbstractProvider
     }
 
     /**
-     * Get listings path
+     * Get page
      *
      * @return  string
      */
-    public function getListingsPath()
+    public function getFrom()
     {
+        if ($this->page) {
+            $from = ($this->page - 1) * $this->count;
+
+            if ($from) {
+                return $from;
+            }
+        }
+
         return null;
     }
 
@@ -129,20 +196,12 @@ class Govt extends AbstractProvider
     }
 
     /**
-     * Get page
+     * Get listings path
      *
      * @return  string
      */
-    public function getFrom()
+    public function getListingsPath()
     {
-        if ($this->page) {
-            $from = ($this->page - 1) * $this->count;
-
-            if ($from) {
-                return $from;
-            }
-        }
-
         return null;
     }
 
@@ -153,22 +212,7 @@ class Govt extends AbstractProvider
      */
     public function getQueryString()
     {
-        $query_params = [
-            'query' => 'getKeyword',
-            'from' => 'getFrom',
-            'size' => 'getCount',
-        ];
-
-        $query_string = [];
-
-        array_walk($query_params, function ($value, $key) use (&$query_string) {
-            $computed_value = $this->$value();
-            if (!is_null($computed_value)) {
-                $query_string[$key] = $computed_value;
-            }
-        });
-
-        return http_build_query($query_string);
+        return http_build_query($this->queryParams);
     }
 
     /**
@@ -190,5 +234,21 @@ class Govt extends AbstractProvider
     public function getVerb()
     {
         return 'GET';
+    }
+
+    /**
+     * Attempts to update current query parameters.
+     *
+     * @param  string  $value
+     * @param  string  $key
+     *
+     * @return Careerbuilder
+     */
+    protected function updateQuery($value, $key)
+    {
+        if (array_key_exists($key, $this->queryParams)) {
+            $this->queryParams[$key] = $value;
+        }
+        return $this;
     }
 }