+++ CHANGELOG.md
@@ -1,6 +1,24 @@
 # Changelog
 All Notable changes to `jobs-govt` will be documented in this file
 
+## 0.6.0 - 2015-09-28
+
+### Added
+- Support for all setter methods outlined in the [Government Jobs API](http://search.digitalgov.gov/developer/jobs.html)
+- Readme documentation for all supported methods
+
+### Deprecated
+- Nothing
+
+### Fixed
+- Nothing
+
+### Removed
+- Nothing
+
+### Security
+- Nothing
+
 ## 0.5.0 - 2015-08-14
 
 ### Added